refactor(ai): Slice B near-cap planning splits and 450 line ratchet
Concern-split declare-war suppression, research funding/decision,
recruitment ledger, treasury regiment bids, and economy cargo helpers
so every AI lib/src file stays under the new 450 physical-line ceiling.
