+++ packages/colonizethis_ai/lib/src/planning/diplomatic_candidate_scoring_declare_war.dart
@@ -1,15 +1,10 @@
-import 'army_conquest_prep.dart' show regimentCountForPlayer;
 import 'diplomatic_candidate_scoring_declare_war_bonuses.dart';
 import 'diplomatic_candidate_scoring_declare_war_context.dart';
+import 'diplomatic_candidate_scoring_declare_war_suppression_adjacent_gp.dart';
+import 'diplomatic_candidate_scoring_declare_war_suppression_war_concentration.dart';
 import 'diplomatic_candidate_scoring_shared.dart';
-import 'expand_phase_planner.dart';
 import 'observer_goal_phase.dart';
 import 'phase_planner_diplomacy_filter.dart';
-import 'planning_helpers.dart'
-    show
-        anyInvadableProvinceOwnedByMinor,
-        gpFactionIdsAtWarWith,
-        isAtWarWithAnyGreatPower;
 import 'planning_imports.dart';
 
 /// Declare-war score ladder.
@@ -45,11 +40,11 @@ int? _declareWarSuppressedScore(
       _declareWarSuppressedColonialLiteScore(ctx) ??
       _declareWarSuppressedExpandColonialScore(ctx) ??
       _declareWarSuppressedStalledOwFrontierScore(ctx) ??
-      _declareWarSuppressedAdjacentGpScore(
+      declareWarSuppressedAdjacentGpScore(
         ctx,
         sameTurnPriorDiplomaticOrders: sameTurnPriorDiplomaticOrders,
       ) ??
-      _declareWarSuppressedWarConcentrationScore(
+      declareWarSuppressedWarConcentrationScore(
         ctx,
         sameTurnPriorDiplomaticOrders: sameTurnPriorDiplomaticOrders,
       ) ??
@@ -223,251 +218,6 @@ int? _declareWarSuppressedStalledOwFrontierScore(DeclareWarTargetContext ctx) {
       !ctx.invadableGpBlockerWeaker &&
       !ctx.invadableGpBlocker) {
     return 0;
-  }
-  return null;
-}
-
-int? _declareWarSuppressedAdjacentGpScore(
-  DeclareWarTargetContext ctx, {
-  Orders? sameTurnPriorDiplomaticOrders,
-}) {
-  if (ctx.order.type == DiplomaticOrderType.declareWar && ctx.isAdjacentGp) {
-    final attackerOw = ctx.snapshot.conquest.oldWorldProvincesOwned;
-    final targetOw = ctx.targetProvinceCount;
-    if (ctx.targetIsGreatPower) {
-      if (regimentCountForPlayer(ctx.game, ctx.nationId) == 0 &&
-          isBelowObserverConquestQuota(attackerOw)) {
-        return 0;
-      }
-      if (ctx.targetNotAlreadyAtWar &&
-          isMutualBelowQuotaPlateauPeer(
-            ownOw: attackerOw,
-            partnerOw: targetOw,
-          ) &&
-          targetOw <= attackerOw + 1) {
-        return 0;
-      }
-      if (isBelowObserverConquestQuota(targetOw) &&
-          isBelowObserverConquestQuota(attackerOw) &&
-          !ctx.invadableGpBlocker &&
-          ctx.targetNotAlreadyAtWar &&
-          attackerOw >= kObserverDefaultStartOldWorldProvincesPerGp &&
-          targetOw <= attackerOw) {
-        return 0;
-      }
-      final minorsOwnInvadable = anyInvadableProvinceOwnedByMinor(
-        game: ctx.game,
-        snapshot: ctx.snapshot,
-        provinceOwner: ctx.provinceOwner,
-      );
-      if (minorsOwnInvadable &&
-          isBelowObserverConquestQuota(attackerOw) &&
-          isBelowObserverConquestQuota(targetOw) &&
-          (targetOw - attackerOw).abs() <= 2 &&
-          !ctx.invadableGpBlocker &&
-          ctx.targetNotAlreadyAtWar) {
-        return 0;
-      }
-      if (attackerOw >= kObserverDefaultStartOldWorldProvincesPerGp &&
-          isBelowObserverConquestQuota(targetOw) &&
-          targetOw <= kObserverDefaultStartOldWorldProvincesPerGp + 1 &&
-          !ctx.invadableGpBlocker) {
-        return 0;
-      }
-      if (isBelowObserverConquestQuota(attackerOw) &&
-          pendingDeclareWarFrom(
-            sameTurnPriorDiplomaticOrders: sameTurnPriorDiplomaticOrders,
-            declarerFactionId: ctx.order.targetFactionId,
-            targetFactionId: ctx.nationId,
-          )) {
-        return 0;
-      }
-      if (isBelowObserverConquestQuota(targetOw) &&
-          targetOw <= kFewOldWorldProvincesDefendThreshold &&
-          !isBelowObserverConquestQuota(attackerOw) &&
-          ctx.targetNotAlreadyAtWar) {
-        return 0;
-      }
-      if (!ctx.invadableGpBlocker &&
-          isBelowObserverConquestQuota(targetOw) &&
-          regimentCountForPlayer(ctx.game, ctx.order.targetFactionId) == 0 &&
-          ctx.targetNotAlreadyAtWar) {
-        return 0;
-      }
-      if (isBelowObserverConquestQuota(targetOw) &&
-          !ctx.invadableGpBlocker &&
-          ctx.targetNotAlreadyAtWar &&
-          ((!isBelowObserverConquestQuota(attackerOw)) ||
-              (ctx.currentTurn <= kDeclareWarEarlyAntiDogpileMaxTurn &&
-                  attackerOw > targetOw))) {
-        return 0;
-      }
-      final belowQuotaSuppressLead =
-          targetOw <= kFewOldWorldProvincesDefendThreshold
-          ? 1
-          : kUnwinnableSoleGpMinProvinceDeficit;
-      if (isBelowObserverConquestQuota(targetOw) &&
-          !ctx.invadableGpBlocker &&
-          attackerOw >= targetOw + belowQuotaSuppressLead) {
-        return 0;
-      }
-      if (targetOw <= kObserverDefaultStartOldWorldProvincesPerGp &&
-          attackerOw > targetOw &&
-          !ctx.invadableGpBlocker &&
-          ctx.targetNotAlreadyAtWar) {
-        return 0;
-      }
-      if (targetOw <= kObserverDefaultStartOldWorldProvincesPerGp &&
-          attackerOw >= kObserverDefaultStartOldWorldProvincesPerGp + 1 &&
-          !ctx.invadableGpBlocker) {
-        return 0;
-      }
-      if (isBelowObserverConquestQuota(targetOw) &&
-          targetOw <= kObserverDefaultStartOldWorldProvincesPerGp &&
-          !isBelowObserverConquestQuota(attackerOw) &&
-          !ctx.invadableGpBlocker &&
-          ctx.targetNotAlreadyAtWar) {
-        return 0;
-      }
-      if (isBelowObserverConquestQuota(targetOw) &&
-          targetOw <= kObserverDefaultStartOldWorldProvincesPerGp &&
-          attackerOw > targetOw &&
-          !ctx.invadableGpBlocker &&
-          ctx.targetNotAlreadyAtWar) {
-        return 0;
-      }
-      if (!isBelowObserverConquestQuota(attackerOw) &&
-          isBelowObserverConquestQuota(targetOw) &&
-          targetOw <= kStalledOldWorldProvinceThreshold &&
-          !ctx.invadableGpBlocker &&
-          ctx.targetNotAlreadyAtWar) {
-        return 0;
-      }
-      if (isBelowObserverConquestQuota(attackerOw) &&
-          targetOw >= attackerOw + kUnwinnableSoleGpMinProvinceDeficit) {
-        return 0;
-      }
-      if (isBelowObserverConquestQuota(attackerOw) &&
-          attackerOw <= kObserverDefaultStartOldWorldProvincesPerGp + 1 &&
-          ctx.isAdjacentGp &&
-          !ctx.invadableGpBlocker &&
-          ctx.targetIsInvadableOwner &&
-          targetOw > attackerOw) {
-        return 0;
-      }
-    }
-    if (!ctx.invadableGpBlocker &&
-        attackerOw <= kFewOldWorldProvincesDefendThreshold &&
-        targetOw > attackerOw) {
-      return 0;
-    }
-    if (!ctx.invadableGpBlocker &&
-        attackerOw <= kFewOldWorldProvincesDefendThreshold &&
-        ctx.snapshot.conquest.invadableProvinceIdsSorted.isNotEmpty &&
-        ctx.targetNotAlreadyAtWar) {
-      return 0;
-    }
-  }
-  if (ctx.order.type == DiplomaticOrderType.declareWar &&
-      ctx.isAdjacentGp &&
-      ctx.targetIsGreatPower &&
-      ctx.stalledOwExpansion) {
-    final targetOw = ctx.targetProvinceCount;
-    if (!ctx.invadableGpBlocker &&
-        targetOw <= kFewOldWorldProvincesDefendThreshold &&
-        ctx.snapshot.conquest.oldWorldProvincesOwned >=
-            targetOw + kDeclareWarAggressorSuppressWeakGpLeadThreshold) {
-      return 0;
-    }
-  }
-  return null;
-}
-
-int? _declareWarSuppressedWarConcentrationScore(
-  DeclareWarTargetContext ctx, {
-  Orders? sameTurnPriorDiplomaticOrders,
-}) {
-  final atWarWithGp = isAtWarWithAnyGreatPower(ctx.game, ctx.snapshot);
-  if (ctx.stalledOwExpansion &&
-      atWarWithGp &&
-      ctx.isAdjacentGp &&
-      ctx.targetIsGreatPower &&
-      ctx.targetNotAlreadyAtWar) {
-    return 0;
-  }
-  if (ctx.isAdjacentGp && ctx.targetIsGreatPower && ctx.targetNotAlreadyAtWar) {
-    final attackerGpWarCount = gpFactionIdsAtWarWith(
-      ctx.game,
-      ctx.snapshot,
-    ).length;
-    if (attackerGpWarCount >= 2) {
-      return 0;
-    }
-    final targetGpId = ctx.order.targetFactionId;
-    final targetOw = ctx.targetProvinceCount;
-    final targetGpWarCount = greatPowerWarCountOnTarget(
-      game: ctx.game,
-      targetGpId: targetGpId,
-      sameTurnPriorDiplomaticOrders: sameTurnPriorDiplomaticOrders,
-    );
-    if (targetGpWarCount >= 2) {
-      return 0;
-    }
-    final attackerOw = ctx.snapshot.conquest.oldWorldProvincesOwned;
-    if (isBelowObserverConquestQuota(targetOw) && targetGpWarCount >= 1) {
-      return 0;
-    }
-    if (isBelowObserverConquestQuota(targetOw) &&
-        attackerOw >= targetOw + 2 &&
-        !ctx.invadableGpBlocker) {
-      return 0;
-    }
-  }
-  // While an invadable OW frontier has a GP blocker, do not open (or stack)
-  // wars on other adjacent GPs — applies above the stalled OW band (seed-42 gp4).
-  final primaryGpBlocker = primaryInvadableOldWorldGpBlocker(
-    game: ctx.game,
-    snapshot: ctx.snapshot,
-  );
-  if (atWarWithGp &&
-      ctx.isAdjacentGp &&
-      ctx.targetIsGreatPower &&
-      ctx.targetNotAlreadyAtWar &&
-      ctx.snapshot.conquest.invadableProvinceIdsSorted.isNotEmpty &&
-      primaryGpBlocker != null &&
-      ctx.order.targetFactionId != primaryGpBlocker) {
-    return 0;
-  }
-  if (ctx.stalledOwExpansion &&
-      ctx.invadableGpBlocker &&
-      ctx.targetProvinceCount > ctx.snapshot.conquest.oldWorldProvincesOwned &&
-      ctx.hasInvadableMinorOwner) {
-    return 0;
-  }
-  if (ctx.stalledOwExpansion &&
-      ctx.isTribeTarget &&
-      !ctx.tribeOwnsOwInvadable &&
-      !(ctx.colonialPressure && ctx.ownsInvadableNw) &&
-      (ctx.behindVictoryPace ||
-          ctx.hasInvadableMinorOwner ||
-          ctx.atWarInvadableOwMinor)) {
-    return kDeclareWarNonAdjacentSuppressedScore;
-  }
-  if (ctx.suppressGpDeclareWar &&
-      ctx.isAdjacentGp &&
-      !ctx.invadableGpBlocker &&
-      !(ctx.stalledOwExpansion && ctx.invadableGpBlockerWeaker)) {
-    return kDeclareWarNonAdjacentSuppressedScore;
-  }
-  if (ctx.suppressGpDeclareWar &&
-      ctx.isAdjacentGp &&
-      ctx.stalledOwExpansion &&
-      ctx.behindVictoryPace &&
-      ctx.hasInvadableMinorOwner &&
-      !ctx.invadableGpBlocker &&
-      !ctx.invadableGpBlockerWeaker &&
-      ctx.lowWarLikelihood) {
-    return kDeclareWarNonAdjacentSuppressedScore;
   }
   return null;
 }
